refactor: introduce `Bookmark` component
Instead of separate `BookmarkFolder` and `BookmarkLink` components.
